Nested friendslist component to be side by side with chatroom (#168)
* Nested friendslist component to be side by side with chatroom

* Removed CSS from the global file

* Got the auth2 tests working again

* Merged main into the current branch. Added the chat vulgarity censor to the chatUtils sendMessage function

* Added parent useState to keep track of currently-selected patients. Fixed up CSS structure in FriendsList. Added dummy function to simulate a call to open a new private chatroom. Currently only logs to the console.
